update plugin manager.

+++ src/engine/PluginManager.js
@@ -5,10 +5,24 @@ const fs = require("fs");
 const os = require("os");
 const path = require("path");
 const request = require("request");
+const request_promise = require("request-promise");
 const progress = require("request-progress");
 
 let localBoardName = "";
 let localPlugins = {};
+
+const decodeArduinoLibraryConfig = function(targetFile){
+  let tmpData = fs.readFileSync(targetFile,"utf8");
+  let lines = tmpData.split('\n');
+  let tdata = lines.map(l => l.split("="));
+  let res = {};
+  tdata.forEach(el=>{
+    if(el.length && el.length === 2) {
+      res[el[0].trim()] = el[1].trim();
+    }
+  });
+  return res;
+};
 
 const listPlugin = function(dir) {
   let plugins = {};
@@ -115,7 +129,7 @@ const listKidBrightPlugin = function(dir) {
   return plugins;
 };
 
-let listExamples = function(exampleDir) {
+const listExamples = function(exampleDir) {
   let exampleInfo = [];
   if (fs.existsSync(exampleDir)) {
     let exampleFolders = fs.readdirSync(exampleDir);
@@ -133,65 +147,121 @@ let listExamples = function(exampleDir) {
   }
   return exampleInfo;
 };
-var listCategoryPlugins = function(pluginDir,boardInfo) {
-  var categories = [];
-  var allPlugin = {};
+const listCategoryPlugins = function(pluginDir,boardInfo) {
+  let categories = [];
+  let allPlugin = {};
   if (fs.existsSync(pluginDir)) {
-    var cats = fs.readdirSync(pluginDir);
+    let cats = fs.readdirSync(pluginDir);
     cats.forEach(cat => {
       let dir = `${pluginDir}/${cat}`;
-      let infoFile = `${dir}/library.json`;
+      if(!fs.lstatSync(dir).isDirectory()){
+        return;
+      }
+      //------- load kidbright plugins -------//
       let kbPluginInfoFile = `${dir}/${cat}.json`;
-      let srcDir = `${dir}/src`;
-      let blockDir = `${dir}/blocks`;
-      let exampleDir = `${dir}/examples`;
-      if (fs.existsSync(infoFile) && fs.existsSync(srcDir) &&
-          fs.existsSync(blockDir)) {
-        let pluginInfo = JSON.parse(fs.readFileSync(infoFile, "utf8"));
-        //----- check correct platform -----//
-        if(boardInfo !== undefined){
-          if(typeof(pluginInfo.platform) === "string"
+      if (fs.existsSync(kbPluginInfoFile)) {
+        let catInfoFile = JSON.parse(fs.readFileSync(kbPluginInfoFile,"utf8"));
+        let blockPlugins = listKidBrightPlugin(dir);
+        categories.push({
+                          directory: dir,
+                          dirName: cat,
+                          plugins: blockPlugins,
+                          category: catInfoFile,
+                        });
+        Object.assign(allPlugin, blockPlugins);
+      //------- load normal plugins ---------//
+      }else{
+        let pluginInfo = null;
+
+        let infoFile = `${dir}/library.json`;
+        let arduinoInfoFile = `${dir}/library.properties`;
+        if(fs.existsSync(infoFile)) {
+          pluginInfo = JSON.parse(fs.readFileSync(infoFile, "utf8"));
+          if("frameworks" in pluginInfo && "platforms" in pluginInfo) { //this is platformIO config file
+             if(pluginInfo.frameworks === "arduino") {
+                if(pluginInfo.platforms === "*") {
+                  pluginInfo.platform = ["*"];
+                }
+                if(typeof pluginInfo.platforms === "object" && pluginInfo.platforms.includes("espressif32")) {
+                  pluginInfo.platform = ["arduino-esp32"];
+                }
+             }
+             if(pluginInfo.repository && pluginInfo.repository.type === "git") {
+                pluginInfo["git"] = pluginInfo.repository.url.replace(".git","/").trim();
+             }else {
+                return; //this plugin cannot update
+             }
+             if(pluginInfo.authors) {
+               pluginInfo["author"] = pluginInfo.authors.name;
+             }
+             pluginInfo["title"] = pluginInfo["name"];
+             pluginInfo["name"] = pluginInfo["name"].replace(/\s/g,"-").trim();
+          }
+        }else if(fs.existsSync(arduinoInfoFile)) {
+          pluginInfo = decodeArduinoLibraryConfig(arduinoInfoFile);
+          if("url" in pluginInfo) {
+            pluginInfo["git"] = pluginInfo.url;
+          }else {
+            return; //reject
+          }
+          pluginInfo.title = pluginInfo.name;
+          pluginInfo.name = pluginInfo.name.replace(/\s/g,"-").trim();
+          pluginInfo.description = pluginInfo.sentence;
+          pluginInfo.platform = pluginInfo.architectures;
+        }else {
+          return; //there are no info file in this library.
+        }
+        //---------- rejection other board platform ----------//
+        if(boardInfo !== undefined && boardInfo.platform){
+          if(typeof(pluginInfo.platform) === "string" //single param platform
               && !pluginInfo.platform.includes(",")
-              && pluginInfo.platform !== boardInfo.platform){
+              && pluginInfo.platform !== boardInfo.platform
+              && pluginInfo.platform !== "*"){ //any platform ? is that real!?
             return;
-          }else if(typeof(pluginInfo.platform) === "string"
+          }else if(typeof(pluginInfo.platform) === "string" //string with comma
               && pluginInfo.platform.includes(",")){
             let supportedPlugins = pluginInfo.platform.split(",").map(el=>el.trim());
             if(!supportedPlugins.includes(boardInfo.platform)){
               return;
             }
-          }else if(typeof(pluginInfo.platform) === pluginInfo.constructor === Array){
+          }else if(pluginInfo.platform.constructor === Array){ //array of platform
             if(!pluginInfo.platform.includes(boardInfo.platform)){
               return;
             }
           }
         }
-        //---------------------//
-        let plugins = listPlugin(blockDir);
-        let srcFile = fs.readdirSync(srcDir);
+        //---------- load plugin ----------//
+        let srcDir = `${dir}/src`;
+        let blockDir = `${dir}/blocks`;
+        let exampleDir = `${dir}/examples`;
+
+        let blockPlugins = {};
+        let srcFile = [];
+        let srcIncDir = dir;
+        if(fs.existsSync(blockDir)){
+          blockPlugins = listPlugin(blockDir);
+        }
+        if(fs.existsSync(srcDir)){
+          srcFile = fs.readdirSync(srcDir);
+          srcIncDir = srcDir;
+        }else if(fs.readdirSync(dir).find(el=>el.endsWith(".h"))){
+          srcFile = fs.readdirSync(dir);
+          srcIncDir = dir;
+        }
         let exampleInfo = [];
         if (fs.existsSync(exampleDir)) {
           exampleInfo = listExamples(exampleDir);
         }
         categories.push({
-                          directory: dir,
-                          dirName: cat,
-                          sourceFile: srcFile,
-                          plugins: plugins,
-                          category: pluginInfo,
-                          examples: exampleInfo,
-                        });
-        Object.assign(allPlugin, plugins);
-      } else if (fs.existsSync(dir) && fs.existsSync(kbPluginInfoFile)) {
-        let catInfoFile = JSON.parse(fs.readFileSync(kbPluginInfoFile));
-        let plugins = listKidBrightPlugin(dir);
-        categories.push({
-                          directory: dir,
-                          dirName: cat,
-                          plugins: plugins,
-                          category: catInfoFile,
-                        });
-        Object.assign(allPlugin, plugins);
+                            directory: dir,
+                            dirName: cat,
+                            sourceFile: srcFile,
+                            sourceIncludeDir : srcIncDir,
+                            plugins: blockPlugins,
+                            category: pluginInfo,
+                            examples: exampleInfo,
+                          });
+        Object.assign(allPlugin, blockPlugins);
       }
     });
   }
@@ -199,7 +269,7 @@ var listCategoryPlugins = function(pluginDir,boardInfo) {
 };
 //TODO : look for platform blocks
 
-var loadPlugin = function(boardInfo) {
+const loadPlugin = function(boardInfo) {
   if ((Object.entries(localPlugins).length === 0 && localPlugins.constructor ===
       Object) || (boardInfo.name !== localBoardName)) { // check empty object !!!
     //load mother platform
@@ -216,7 +286,7 @@ var loadPlugin = function(boardInfo) {
     let globalPlugins = listCategoryPlugins(
         util.pluginDir,
         boardInfo
-    )
+    );
     //join all together
     let allPlugins = {};
     Object.assign(allPlugins, platformPlugins.plugins);
@@ -231,15 +301,15 @@ var loadPlugin = function(boardInfo) {
   return localPlugins;
 };
 
-var clearListedPlugin = function() {
+const clearListedPlugin = function() {
   localPlugins = {};
 };
 
-var plugins = function(boardInfo) {
+const plugins = function(boardInfo) {
   let lpg = loadPlugin(boardInfo);
   return lpg.categories;
 };
-var performPluginSearch = function(name,queryMode, value, start = 0) {
+const performPluginSearch = function(name,queryMode, value, start = 0) {
   return new Promise((resolve, reject) => {
     let onlinePlugins = [];
     Vue.prototype.$db.collection("plugins").
@@ -249,7 +319,7 @@ var performPluginSearch = function(name,queryMode, value, start = 0) {
     limit(50).
     get().
     then(data => {
-      var lastVisible = data.docs[data.docs.length - 1];
+      const lastVisible = data.docs[data.docs.length - 1];
       data.forEach(element => {
         onlinePlugins.push(element.data());
       });
@@ -260,22 +330,21 @@ var performPluginSearch = function(name,queryMode, value, start = 0) {
     });
   });
 };
-var performPluginNameSearch = function(name, column, value, start = 0) {
+const performPluginNameSearch = function(name, column, value, start = 0) {
   return new Promise((resolve, reject) => {
     let onlinePlugins = [];
-    var strSearch = name;
-    var strlength = strSearch.length;
-    var strFrontCode = strSearch.slice(0, strlength - 1);
-    var strEndCode = strSearch.slice(strlength - 1, strSearch.length);
+    let strSearch = name;
+    let strlength = strSearch.length;
+    let strFrontCode = strSearch.slice(0, strlength - 1);
+    let strEndCode = strSearch.slice(strlength - 1, strSearch.length);
 
-    var startcode = strSearch;
-    var endcode = strFrontCode +
-        String.fromCharCode(strEndCode.charCodeAt(0) + 1);
+    let startcode = strSearch;
+    let endcode = strFrontCode + String.fromCharCode(strEndCode.charCodeAt(0) + 1);
     Vue.prototype.$db.collection("plugins").where("name", ">=", startcode) //search start with
     .where("name", "<", endcode).where(column, "==", value).orderBy("name")
     //.startAfter(start)
     .limit(50).get().then(data => {
-      var lastVisible = data.docs[data.docs.length - 1];
+      let lastVisible = data.docs[data.docs.length - 1];
       data.forEach(element => {
         onlinePlugins.push(element.data());
       });
@@ -285,10 +354,10 @@ var performPluginNameSearch = function(name, column, value, start = 0) {
     });
   });
 };
-var listOnlinePlugin = function(boardInfo, name = "", start = 0) {
+const listOnlinePlugin = function(boardInfo, name = "", start = 0) {
   return new Promise((resolve, reject) => {
     let onlinePlugins = [];
-    if (name == "") { //list all
+    if (name === "") { //list all
       performPluginSearch("board","==", boardInfo.name).then(res => {
         onlinePlugins = onlinePlugins.concat(res.plugins);
         return performPluginSearch("platform","array-contains",boardInfo.platform);
@@ -308,12 +377,11 @@ var listOnlinePlugin = function(boardInfo, name = "", start = 0) {
       }).catch(err => {
         reject(err);
       });
-
     }
   });
 };
 
-var installPluginByName = function(name, cb) {
+const installPluginByName = function(name, cb) {
   return new Promise((resolve, reject) => {
     Vue.prototype.$db.collection("plugins").
     where("name", "==", name).
@@ -330,19 +398,13 @@ var installPluginByName = function(name, cb) {
     return installOnlinePlugin(info, cb);
   });
 };
-var installOnlinePlugin = function(info, cb) {
-  if (info.board) {
-    var targetDir = util.boardDir + "/" + info.board + "/plugin";
-  } else if (info.platform) {
-    var targetDir = util.platformDir + "/" + info.platform + "/plugin";
-  } else {
-    throw "no target defined";
-  }
+const installOnlinePlugin = function(info, cb) {
+  let targetDir = util.pluginDir;
   return new Promise((resolve, reject) => { //download zip
     if (!info.git) { reject("no git found"); }
-    var zipUrl = info.git + "/archive/master.zip";
-    var zipFile = os.tmpdir() + "/" + util.randomString(10) + ".zip";
-    var file = fs.createWriteStream(zipFile);
+    let zipUrl = info.git + "/archive/master.zip";
+    let zipFile = os.tmpdir() + "/" + util.randomString(10) + ".zip";
+    let file = fs.createWriteStream(zipFile);
     progress(
         request(zipUrl),
         {
@@ -352,7 +414,7 @@ var installOnlinePlugin = function(info, cb) {
           follow: true,
         },
     ).on("progress", function(state) {
-      cb & cb({process: "board", status: "DOWNLOAD", state: state});
+      cb && cb({process: "board", status: "DOWNLOAD", state: state});
     }).on("error", function(err) {
       reject(err);
     }).on("end", function() {
@@ -361,11 +423,11 @@ var installOnlinePlugin = function(info, cb) {
     }).pipe(file);
   }).then((zipFile) => { //unzip file
     return util.unzip(zipFile, {dir: targetDir}, p => {
-      cb & cb({process: "board", status: "UNZIP", state: p});
+      cb && cb({process: "board", status: "UNZIP", state: p});
     });
   }).then(() => { //rename folder
     //rename ended with word '-master' in boards
-    var dirs = fs.readdirSync(targetDir);
+    let dirs = fs.readdirSync(targetDir);
     for (let i = 0; i < dirs.length; i++) {
       let dirname = path.join(targetDir, dirs[i]);
       if (fs.lstatSync(dirname).isDirectory() && dirname.endsWith("-master")) {
@@ -378,7 +440,7 @@ var installOnlinePlugin = function(info, cb) {
   });
 };
 
-var removePlugin = function(pluginInfo, isBackupRemove = false) {
+const removePlugin = function(pluginInfo, isBackupRemove = false) {
   return new Promise((resolve, reject) => {
     let target = pluginInfo.directory;
     if (isBackupRemove) {
@@ -393,35 +455,67 @@ var removePlugin = function(pluginInfo, isBackupRemove = false) {
   });
 };
 
-var backupPlugin = function(pluginInfo) {
-  if (pluginInfo.board) {
-    var targetDir = util.boardDir + "/" + pluginInfo.board + "/plugin";
-  } else if (pluginInfo.platform) {
-    var targetDir = util.platformDir + "/" + pluginInfo.platform + "/plugin";
-  } else {
-    throw "no target defined";
-  }
+const backupPlugin = function(pluginInfo) {
   return new Promise((resolve, reject) => {
-    let target = `${targetDir}/${pluginInfo.name}`;
-    let newer = `${targetDir}/${pluginInfo.name}-backup-plugin`;
-    fs.renameSync(target, newer);
-    resolve();
+    let target = pluginInfo.directory;
+    if(target.endsWith("/")){
+      target = target.substring(0,target.length - 1);
+    }
+    let newer = `${target}-backup-plugin`;
+    if (!fs.existsSync(target)) {
+      reject("no directory");
+    }else{
+      fs.renameSync(target, newer);
+      resolve();
+    }
   });
 };
 
-var restorePlugin = function(pluginInfo) {
-  if (pluginInfo.board) {
-    var targetDir = util.boardDir + "/" + pluginInfo.board + "/plugin";
-  } else if (pluginInfo.platform) {
-    var targetDir = util.platformDir + "/" + pluginInfo.platform + "/plugin";
-  } else {
-    throw "no target defined";
+const restorePlugin = function(pluginInfo) {
+  let target = pluginInfo.directory;
+  if(target.endsWith("/")){
+    target = target.substring(0,target.length - 1);
   }
-  return new Promise((resolve, reject) => {
-    let target = `${targetDir}/${pluginInfo.name}`;
-    let newer = `${targetDir}/${pluginInfo.name}-backup-plugin`;
-    fs.renameSync(newer, target);
-    resolve();
+  let newer = `${target}-backup-plugin`;
+  fs.renameSync(newer, target);
+  resolve();
+};
+
+const publishPlugin = function(url){
+  return new Promise((resolve,reject)=>{
+    let json = null;
+    if (util.regex.isValidGithubUrl(url)) {
+      reject("wrong github url format");
+      return;
+    }
+    request_promise(url + "raw/master/library.json?random=" + util.randomString()) //add randomstring prevent cached response
+    .then(res => {
+      json = JSON.parse(res);
+      if(json.name) { //search if existing
+        return Vue.prototype.$db.collection("plugins").where("name", "==", json.name).get();
+      } else {
+        return false;
+      }
+    }).then(res => {
+      if (res && res.size >= 1) {
+        return json.version > res.docs[0].data().version;
+      } else {
+        return true;
+      }
+    }).then(res => {
+      if (res) {
+        Vue.prototype.$db.collection("plugins").doc(json.name).set(json);
+        if (res) {
+          resolve("submit your plugin success, please refresh again");
+        }
+      } else {
+        reject("Existing plugin name or is not newest version");
+      }
+    }).catch(err => {
+      console.log("Plugin Publishing Error : ");
+      console.log(err);
+      reject(err);
+    });
   });
 };
 
@@ -435,4 +529,5 @@ export default {
   removePlugin,
   backupPlugin,
   restorePlugin,
+  publishPlugin
 };
